requirements.txt
# ============================================================
# Nepal Electricity Analytics Pipeline — Requirements
# Python 3.10+
# ============================================================

# Core data processing
pandas==2.1.4
numpy==1.26.4
scipy==1.12.0

# Database
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
alembic==1.13.1

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.1

# HTTP / API
requests==2.31.0
urllib3==2.2.0

# Data validation
great-expectations==0.18.12
pydantic==2.6.0

# Scheduling (Apache Airflow)
apache-airflow==2.9.3
apache-airflow-providers-postgres==5.10.0

# Excel / CSV export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Visualization (analytics scripts)
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Reporting
reportlab==4.1.0
jinja2==3.1.3
weasyprint==61.2

# Monitoring
prometheus-client==0.20.0

# Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Type checking / code quality
mypy==1.8.0
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Utilities
tqdm==4.66.1
pytz==2024.1
python-dateutil==2.9.0
tabulate==0.9.0
